pyproject.toml
[tool.poetry]
name = "bot_for_ann"
version = "0.1.0"
description = ""
authors = ["Tony Hat <54199578+t0nyhat@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "20.0a4"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
pep8-naming = "^0.13.2"
flake8-broken-line = "^0.5.0"
flake8-return = "^1.1.3"
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
